backend/service-Inteligence/requirements.txt
fastapi
uvicorn
pika
requests
pdfminer.six
spacy
python-dotenv
transformers
torch
sentencepiece